+++ test/bigtest/interactors/filters/index.js
@@ -1,0 +1,2 @@
+export { default as MultiSelectFilterInteractor } from './multiselect-filter';
+export { default as CheckboxFilterInteractor } from './checkbox-filter';