+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * Add filters for staffSuppress and discoverySuppress. Completes UIIN-289.
 * Display empty elements with the dashes. Refs UIIN-826.
 * Filter item records by material type. Part of UIIN-777.
+* Filter item records by item status. Part of UIIN-771.
 
 ## [1.12.1](https://github.com/folio-org/ui-inventory/tree/v1.12.1) (2019-09-26)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.12.0...v1.12.1)